+++ lib/src/message_framing_state.dart
