+++ lib/src/message_framing_state.freezed.dart
